@@ -33,11 +33,10 @@ struct Config
 
 	struct
 	{
+		u32 anisotropy;
 		u32 maxAnisotropy;
-		f32 maxAnisotropyF;
 		u32 bilinearMode;
 		u32 enableHalosRemoval;
-		u32 screenShotFormat;
 	} texture;
 
 	enum TexrectCorrectionMode {
@@ -66,6 +65,7 @@ struct Config
 		u32 enableShadersStorage;
 		u32 enableLegacyBlending;
 		u32 enableHybridFilter;
+		u32 enableInaccurateTextureCoordinates;
 		u32 enableFragmentDepthWrite;
 		u32 hacks;
 #if defined(OS_ANDROID) || defined(OS_IOS)
@@ -176,6 +176,9 @@ struct Config
 
 		u32 txEnhancedTextureFileStorage;	// Use file storage instead of memory cache for enhanced textures.
 		u32 txHiresTextureFileStorage;		// Use file storage instead of memory cache for hires textures.
+		u32 txNoTextureFileStorage;			// Use no file storage or cache for hires textures.
+
+		u32 txHiresVramLimit; // Limit of uploading hi-res textures to VRAM (in MB)
 
 		wchar_t txPath[PLUGIN_PATH_SIZE]; // Path to texture packs
 		wchar_t txCachePath[PLUGIN_PATH_SIZE]; // Path to store texture cache, that is .htc files
@@ -235,6 +238,7 @@ struct Config
 	};
 
 	struct {
+		u8 enabledKeys[hkTotal];
 		u8 keys[hkTotal];
 	} hotkeys;
 
@@ -245,6 +249,7 @@ struct Config
 	void resetToDefaults();
 	void validate();
 	static const char* hotkeyIniName(u32 _idx);
+	static const char* enabledHotkeyIniName(u32 _idx);
 };
 
 #define hack_Ogre64					(1<<0)  //Ogre Battle 64 background copy
@@ -269,6 +274,7 @@ struct Config
 #define hack_ZeldaMonochrome		(1<<20) //Hack for Zeldas monochrome effects.
 #define hack_TonyHawk				(1<<21) //Hack for Tony Hawk blend mode.
 #define hack_WCWNitro				(1<<22) //Hack for WCW Nitro backgrounds.
+#define hack_fbTextureOffset		(1<<23) //Hack to offset Conker's shadow in CBFD and Bob-ombs in Mario Tennis.
 
 extern Config config;
 
