@@ -1,7 +1,7 @@
 #include <stdarg.h>
 #include <string.h>
 #include <cstdlib>
-#include "GLideNHQ/Ext_TxFilter.h"
+#include "GLideNHQ/TxFilterExport.h"
 #include <Graphics/Context.h>
 #include <Graphics/Parameters.h>
 
