@@ -24,38 +24,29 @@
 #ifndef __TXHIRESCACHE_H__
 #define __TXHIRESCACHE_H__
 
-/* support hires textures
- *   0: disable
- *   1: enable
- */
-#define HIRES_TEXTURE 1
-
 #include "TxCache.h"
 #include "TxQuantize.h"
 #include "TxImage.h"
 #include "TxReSample.h"
+#include "TxHiResLoader.h"
 
-class TxHiResCache : public TxCache
+class TxHiResCache : public TxCache, public TxHiResLoader
 {
 private:
-  int _maxwidth;
-  int _maxheight;
-  int _maxbpp;
 	bool _abortLoad;
 	bool _cacheDumped;
-	std::unique_ptr<TxImage> _txImage;
-	std::unique_ptr<TxQuantize> _txQuantize;
-	std::unique_ptr<TxReSample> _txReSample;
+
   tx_wstring _texPackPath;
   enum LoadResult {
 	  resOk,
 	  resNotFound,
 	  resError
   };
-  LoadResult loadHiResTextures(const wchar_t * dir_path, boolean replace);
+  LoadResult _loadHiResTextures(const wchar_t * dir_path, boolean replace);
   boolean _HiResTexPackPathExists() const;
 	tx_wstring _getFileName() const override;
 	int _getConfig() const override;
+  bool _load(boolean replace);
 
 public:
   ~TxHiResCache();
@@ -67,8 +58,11 @@ public:
 			   const wchar_t *texPackPath,
 			   const wchar_t *ident,
 			   dispInfoFuncExt callback);
-  bool load(boolean replace);
-  void dump();
+  bool empty() const override;
+  bool add(Checksum checksum, GHQTexInfo *info, int dataSize = 0) override;
+  bool get(Checksum checksum, GHQTexInfo *info) override;
+  bool reload() override;
+  void dump() override;
 };
 
 #endif /* __TXHIRESCACHE_H__ */
