@@ -16,7 +16,7 @@ Q_IMPORT_PLUGIN(QICOPlugin)
 //#define RUN_DIALOG_IN_THREAD
 
 static
-int openConfigDialog(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, bool & _accepted)
+int openConfigDialog(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, unsigned int _maxAnisotropy, bool & _accepted)
 {
 	std::string IniFolder;
 	uint32_t slength = WideCharToMultiByte(CP_ACP, 0, _strFileName, -1, NULL, 0, NULL, NULL);
@@ -34,27 +34,28 @@ int openConfigDialog(const wchar_t * _strFileName, const char * _romName, unsign
 	Dlg.setIniPath(IniFolder.c_str());
 	Dlg.setRomName(_romName);
 	Dlg.setMSAALevel(_maxMSAALevel);
+	Dlg.setMaxAnisotropy(_maxAnisotropy);
 	Dlg.DoModal();
 	_accepted = Dlg.Saved();
 	return 0;
 }
 
-bool runConfigThread(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel)
+bool runConfigThread(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, unsigned int _maxAnisotropy)
 {
 	bool accepted = false;
 #ifdef RUN_DIALOG_IN_THREAD
 	std::thread configThread(openConfigDialog, _strFileName, _maxMSAALevel, std::ref(accepted));
 	configThread.join();
 #else
-	openConfigDialog(_strFileName, _romName, _maxMSAALevel, accepted);
+	openConfigDialog(_strFileName, _romName, _maxMSAALevel, _maxAnisotropy, accepted);
 #endif
 	return accepted;
 
 }
 
-EXPORT bool CALL RunConfig(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel)
+EXPORT bool CALL RunConfig(const wchar_t * _strFileName, const char * _romName, unsigned int _maxMSAALevel, unsigned int _maxAnisotropy)
 {
-	return runConfigThread(_strFileName, _romName, _maxMSAALevel);
+	return runConfigThread(_strFileName, _romName, _maxMSAALevel, _maxAnisotropy);
 }
 
 EXPORT int CALL RunAbout(const wchar_t * _strFileName)
