@@ -35,10 +35,10 @@ void _loadSettings(GlSettings & settings)
 	settings.endGroup();
 
 	settings.beginGroup("texture");
+	config.texture.anisotropy = settings.value("anisotropy", config.texture.anisotropy).toInt();
 	config.texture.maxAnisotropy = settings.value("maxAnisotropy", config.texture.maxAnisotropy).toInt();
 	config.texture.bilinearMode = settings.value("bilinearMode", config.texture.bilinearMode).toInt();
 	config.texture.enableHalosRemoval = settings.value("enableHalosRemoval", config.texture.enableHalosRemoval).toInt();
-	config.texture.screenShotFormat = settings.value("screenShotFormat", config.texture.screenShotFormat).toInt();
 	settings.endGroup();
 
 	settings.beginGroup("generalEmulation");
@@ -220,10 +220,10 @@ void writeSettings(const char * _strIniFolder)
 		settings.endGroup();
 
 		settings.beginGroup("texture");
+		settings.setValue("anisotropy", config.texture.anisotropy);
 		settings.setValue("maxAnisotropy", config.texture.maxAnisotropy);
 		settings.setValue("bilinearMode", config.texture.bilinearMode);
 		settings.setValue("enableHalosRemoval", config.texture.enableHalosRemoval);
-		settings.setValue("screenShotFormat", config.texture.screenShotFormat);
 		settings.endGroup();
 
 		settings.beginGroup("generalEmulation");
@@ -423,10 +423,10 @@ void saveCustomRomSettings(const char * _strIniFolder, const char * _strRomName)
 	settings.endGroup();
 
 	settings.beginGroup("texture");
+	WriteCustomSetting(texture, anisotropy);
 	WriteCustomSetting(texture, maxAnisotropy);
 	WriteCustomSetting(texture, bilinearMode);
 	WriteCustomSetting(texture, enableHalosRemoval);
-	WriteCustomSetting(texture, screenShotFormat);
 	settings.endGroup();
 
 	settings.beginGroup("generalEmulation");
@@ -502,6 +502,7 @@ void saveCustomRomSettings(const char * _strIniFolder, const char * _strRomName)
 	WriteCustomSetting2(onScreenDisplay, showPercent, percent);
 	WriteCustomSetting2(onScreenDisplay, showInternalResolution, internalResolution);
 	WriteCustomSetting2(onScreenDisplay, showRenderingResolution, renderingResolution);
+	WriteCustomSetting2(onScreenDisplay, showStatistics, statistics);
 	WriteCustomSetting2(onScreenDisplay, osdPos, pos);
 	settings.endGroup();
 
