@@ -145,75 +145,77 @@
  # d e f i n e   I D C _ O S D _ C O L O R                                       1 1 1 6  
  # d e f i n e   I D C _ O S D _ P R E V I E W                                   1 1 1 7  
  # d e f i n e   I D C _ C H K _ D U M P _ L O W                                 1 1 1 8  
- # d e f i n e   I D C _ C H K _ D U M P _ N O R M A L                           1 1 1 9  
- # d e f i n e   I D C _ C H K _ D U M P _ D E T A I L                           1 1 2 0  
- # d e f i n e   I D C _ S E R G E Y                                             1 1 2 1  
- # d e f i n e   I D C _ O L I V I E R                                           1 1 2 2  
- # d e f i n e   I D C _ R Y A N                                                 1 1 2 3  
- # d e f i n e   I D C _ R O L E 1                                               1 1 2 4  
- # d e f i n e   I D C _ T X T _ D E V E L O P E R                               1 1 2 5  
- # d e f i n e   I D C _ R O L E 2                                               1 1 2 6  
- # d e f i n e   I D C _ T X T _ B E T A _ T E S T E R                           1 1 2 7  
- # d e f i n e   I D C _ R O L E 3                                               1 1 2 8  
- # d e f i n e   I D C _ T X T _ G U I _ D E S I G N E R                         1 1 2 9  
- # d e f i n e   I D C _ A B O U T _ I C O N                                     1 1 3 0  
- # d e f i n e   I D C _ G A M E _ P R O F I L E                                 1 1 3 1  
- # d e f i n e   I D C _ S A V E _ S E T T I N G S _ S T A T I C                 1 1 3 2  
- # d e f i n e   I D C _ U S E _ P R O F I L E                                   1 1 3 3  
- # d e f i n e   I D C _ S E T T I N G S _ P R O F I L E _ S T A T I C           1 1 3 4  
- # d e f i n e   I D C _ P R O F I L E                                           1 1 3 5  
- # d e f i n e   I D C _ R E M O V E _ P R O F I L E                             1 1 3 6  
- # d e f i n e   I D C _ G A M E _ P R O F I L E _ N A M E                       1 1 3 7  
- # d e f i n e   I D C _ A A _ I N F O                                           1 1 3 8  
- # d e f i n e   I D C _ A A _ G R O U P                                         1 1 3 9  
- # d e f i n e   I D C _ F I L T E R I N G _ G R O U P                           1 1 4 0  
- # d e f i n e   I D C _ V I D E O _ G R O U P                                   1 1 4 1  
- # d e f i n e   I D C _ A N I S O T R O P I C                                   1 1 4 2  
- # d e f i n e   I D C _ B I L I N E A R                                         1 1 4 3  
- # d e f i n e   I D C _ A N I S O T R O P I C _ O F F                           1 1 4 4  
- # d e f i n e   I D C _ A N I S O T R O P I C _ H I G H                         1 1 4 5  
- # d e f i n e   I D C _ A A _ O F F                                             1 1 4 6  
- # d e f i n e   I D C _ A A _ H I G H                                           1 1 4 7  
- # d e f i n e   I D C _ D I T H E R I N G _ G R O U P                           1 1 4 8  
- # d e f i n e   I D C _ C H K _ 5 B I T _ Q U A N T I Z A T I O N               1 1 4 9  
- # d e f i n e   I D C _ C H K _ A P P L Y _ T O _ O U T P U T                   1 1 5 0  
- # d e f i n e   I D C _ C H K _ H I R E S _ N O I S E                           1 1 5 1  
- # d e f i n e   I D C _ C M B _ P A T T E R N                                   1 1 5 2  
- # d e f i n e   I D C _ C M B _ L A N G U A G E                                 1 1 5 3  
- # d e f i n e   I D C _ P A T T E R N                                           1 1 5 4  
- # d e f i n e   I D C _ I N P U T                                               1 1 5 5  
- # d e f i n e   I D C _ I N F O                                                 1 1 5 6  
- # d e f i n e   I D C _ T X T _ F U L L _ S C R E E N _ R E S                   1 1 5 7  
- # d e f i n e   I D C _ T X T _ R E F R E S H _ R A T E                         1 1 5 8  
- # d e f i n e   I D C _ T X T _ W I N D O W E D _ R E S O L U T I O N           1 1 5 9  
- # d e f i n e   I D C _ T X T _ A S P E C T _ R A T I O                         1 1 6 0  
- # d e f i n e   I D C _ L A N G U A G E                                         1 1 6 1  
- # d e f i n e   I D C _ I N T E R N A L _ R E S _ G R O U P                     1 1 6 2  
- # d e f i n e   I D C _ G A M M A _ G R O U P                                   1 1 6 3  
- # d e f i n e   I D C _ 2 D _ E L E M E N T S _ G R O U P                       1 1 6 4  
- # d e f i n e   I D C _ T X T _ F I X B L A C K _ L I N E S                     1 1 6 5  
- # d e f i n e   I D C _ R E N D E R _ 2 D _ E L E M E N T S                     1 1 6 6  
- # d e f i n e   I D C _ T X T _ B A C K G R O U N D _ R E N D E R I N G         1 1 6 7  
- # d e f i n e   I D C _ T X T _ F R A M E B U F F E R _ S W A P                 1 1 6 8  
- # d e f i n e   I D C _ T X T _ C O P Y _ C O L O R _ B U F F E R               1 1 6 9  
- # d e f i n e   I D C _ T X T _ C O P Y _ D E P T H _ B U F F E R               1 1 7 0  
- # d e f i n e   I D C _ T X T _ N 6 4 _ D E P T H _ C O M P A R E               1 1 7 1  
- # d e f i n e   I D C _ F R A M E _ B U F F E R _ I N F O                       1 1 7 2  
- # d e f i n e   I D C _ T X T _ N 6 4 _ T E X T U R E                           1 1 7 3  
- # d e f i n e   I D C _ T X T _ T E X T U R E _ F I L T E R                     1 1 7 4  
- # d e f i n e   I D C _ T X T _ E N H A N C E M E N T                           1 1 7 5  
- # d e f i n e   I D C _ P O S I T I O N                                         1 1 7 6  
- # d e f i n e   I D C _ F O N T                                                 1 1 7 7  
- # d e f i n e   I D C _ T X T _ S I Z E                                         1 1 7 8  
- # d e f i n e   I D C _ P X                                                     1 1 7 9  
- # d e f i n e   I D C _ C O L O R                                               1 1 8 0  
- # d e f i n e   I D C _ T X T _ N E W _ G E N                                   1 1 8 1  
- # d e f i n e   I D C _ T X T _ A U T H O R S                                   1 1 8 2  
- # d e f i n e   I D C _ T X T _ A U T H O R _ G L N 6 4                         1 1 8 3  
- # d e f i n e   I D C _ T X T _                                                 1 1 8 4  
- # d e f i n e   I D C _ T X T _ A U T H O R _ G L E S 2 N 6 4                   1 1 8 5  
- # d e f i n e   I D C _ T X T _ A U T H O R _ G L I D E H Q                     1 1 8 6  
- # d e f i n e   I D C _ T X T _ A U T H O R _ Z 6 4                             1 1 8 7  
+ # d e f i n e   I D C _ R E N D E R I N G _ R E S O L U T I O N 2               1 1 1 9  
+ # d e f i n e   I D C _ S T A T I S T I C S                                     1 1 2 0  
+ # d e f i n e   I D C _ C H K _ D U M P _ N O R M A L                           1 1 2 1  
+ # d e f i n e   I D C _ C H K _ D U M P _ D E T A I L                           1 1 2 2  
+ # d e f i n e   I D C _ S E R G E Y                                             1 1 2 3  
+ # d e f i n e   I D C _ O L I V I E R                                           1 1 2 4  
+ # d e f i n e   I D C _ R Y A N                                                 1 1 2 5  
+ # d e f i n e   I D C _ R O L E 1                                               1 1 2 6  
+ # d e f i n e   I D C _ T X T _ D E V E L O P E R                               1 1 2 7  
+ # d e f i n e   I D C _ R O L E 2                                               1 1 2 8  
+ # d e f i n e   I D C _ T X T _ B E T A _ T E S T E R                           1 1 2 9  
+ # d e f i n e   I D C _ R O L E 3                                               1 1 3 0  
+ # d e f i n e   I D C _ T X T _ G U I _ D E S I G N E R                         1 1 3 1  
+ # d e f i n e   I D C _ A B O U T _ I C O N                                     1 1 3 2  
+ # d e f i n e   I D C _ G A M E _ P R O F I L E                                 1 1 3 3  
+ # d e f i n e   I D C _ S A V E _ S E T T I N G S _ S T A T I C                 1 1 3 4  
+ # d e f i n e   I D C _ U S E _ P R O F I L E                                   1 1 3 5  
+ # d e f i n e   I D C _ S E T T I N G S _ P R O F I L E _ S T A T I C           1 1 3 6  
+ # d e f i n e   I D C _ P R O F I L E                                           1 1 3 7  
+ # d e f i n e   I D C _ R E M O V E _ P R O F I L E                             1 1 3 8  
+ # d e f i n e   I D C _ G A M E _ P R O F I L E _ N A M E                       1 1 3 9  
+ # d e f i n e   I D C _ A A _ I N F O                                           1 1 4 0  
+ # d e f i n e   I D C _ A A _ G R O U P                                         1 1 4 1  
+ # d e f i n e   I D C _ F I L T E R I N G _ G R O U P                           1 1 4 2  
+ # d e f i n e   I D C _ V I D E O _ G R O U P                                   1 1 4 3  
+ # d e f i n e   I D C _ A N I S O T R O P I C                                   1 1 4 4  
+ # d e f i n e   I D C _ B I L I N E A R                                         1 1 4 5  
+ # d e f i n e   I D C _ A N I S O T R O P I C _ O F F                           1 1 4 6  
+ # d e f i n e   I D C _ A N I S O T R O P I C _ H I G H                         1 1 4 7  
+ # d e f i n e   I D C _ A A _ O F F                                             1 1 4 8  
+ # d e f i n e   I D C _ A A _ H I G H                                           1 1 4 9  
+ # d e f i n e   I D C _ D I T H E R I N G _ G R O U P                           1 1 5 0  
+ # d e f i n e   I D C _ C H K _ 5 B I T _ Q U A N T I Z A T I O N               1 1 5 1  
+ # d e f i n e   I D C _ C H K _ A P P L Y _ T O _ O U T P U T                   1 1 5 2  
+ # d e f i n e   I D C _ C H K _ H I R E S _ N O I S E                           1 1 5 3  
+ # d e f i n e   I D C _ C M B _ P A T T E R N                                   1 1 5 4  
+ # d e f i n e   I D C _ C M B _ L A N G U A G E                                 1 1 5 5  
+ # d e f i n e   I D C _ P A T T E R N                                           1 1 5 6  
+ # d e f i n e   I D C _ I N P U T                                               1 1 5 7  
+ # d e f i n e   I D C _ I N F O                                                 1 1 5 8  
+ # d e f i n e   I D C _ T X T _ F U L L _ S C R E E N _ R E S                   1 1 5 9  
+ # d e f i n e   I D C _ T X T _ R E F R E S H _ R A T E                         1 1 6 0  
+ # d e f i n e   I D C _ T X T _ W I N D O W E D _ R E S O L U T I O N           1 1 6 1  
+ # d e f i n e   I D C _ T X T _ A S P E C T _ R A T I O                         1 1 6 2  
+ # d e f i n e   I D C _ L A N G U A G E                                         1 1 6 3  
+ # d e f i n e   I D C _ I N T E R N A L _ R E S _ G R O U P                     1 1 6 4  
+ # d e f i n e   I D C _ G A M M A _ G R O U P                                   1 1 6 5  
+ # d e f i n e   I D C _ 2 D _ E L E M E N T S _ G R O U P                       1 1 6 6  
+ # d e f i n e   I D C _ T X T _ F I X B L A C K _ L I N E S                     1 1 6 7  
+ # d e f i n e   I D C _ R E N D E R _ 2 D _ E L E M E N T S                     1 1 6 8  
+ # d e f i n e   I D C _ T X T _ B A C K G R O U N D _ R E N D E R I N G         1 1 6 9  
+ # d e f i n e   I D C _ T X T _ F R A M E B U F F E R _ S W A P                 1 1 7 0  
+ # d e f i n e   I D C _ T X T _ C O P Y _ C O L O R _ B U F F E R               1 1 7 1  
+ # d e f i n e   I D C _ T X T _ C O P Y _ D E P T H _ B U F F E R               1 1 7 2  
+ # d e f i n e   I D C _ T X T _ N 6 4 _ D E P T H _ C O M P A R E               1 1 7 3  
+ # d e f i n e   I D C _ F R A M E _ B U F F E R _ I N F O                       1 1 7 4  
+ # d e f i n e   I D C _ T X T _ N 6 4 _ T E X T U R E                           1 1 7 5  
+ # d e f i n e   I D C _ T X T _ T E X T U R E _ F I L T E R                     1 1 7 6  
+ # d e f i n e   I D C _ T X T _ E N H A N C E M E N T                           1 1 7 7  
+ # d e f i n e   I D C _ P O S I T I O N                                         1 1 7 8  
+ # d e f i n e   I D C _ F O N T                                                 1 1 7 9  
+ # d e f i n e   I D C _ T X T _ S I Z E                                         1 1 8 0  
+ # d e f i n e   I D C _ P X                                                     1 1 8 1  
+ # d e f i n e   I D C _ C O L O R                                               1 1 8 2  
+ # d e f i n e   I D C _ T X T _ N E W _ G E N                                   1 1 8 3  
+ # d e f i n e   I D C _ T X T _ A U T H O R S                                   1 1 8 4  
+ # d e f i n e   I D C _ T X T _ A U T H O R _ G L N 6 4                         1 1 8 5  
+ # d e f i n e   I D C _ T X T _                                                 1 1 8 6  
+ # d e f i n e   I D C _ T X T _ A U T H O R _ G L E S 2 N 6 4                   1 1 8 7  
+ # d e f i n e   I D C _ T X T _ A U T H O R _ G L I D E H Q                     1 1 8 8  
+ # d e f i n e   I D C _ T X T _ A U T H O R _ Z 6 4                             1 1 8 9  
   
  / /   N e x t   d e f a u l t   v a l u e s   f o r   n e w   o b j e c t s  
  / /    
