@@ -23,7 +23,6 @@ namespace graphics {
 		ShaderProgramBinary,
 		ImageTextures,
 		IntegerTextures,
-		ClipControl,
 		N64DepthWithFbFetchDepth,
 		FramebufferFetchColor,
 		TextureBarrier,
@@ -147,6 +146,8 @@ namespace graphics {
 		s32 getTextureUnpackAlignment() const;
 
 		s32 getMaxTextureSize() const;
+
+		f32 getMaxAnisotropy() const;
 
 		struct BindImageTextureParameters {
 			ImageUnitParam imageUnit;
@@ -298,7 +299,6 @@ namespace graphics {
 		static bool ShaderProgramBinary;
 		static bool ImageTextures;
 		static bool IntegerTextures;
-		static bool ClipControl;
 		static bool FramebufferFetchDepth;
 		static bool FramebufferFetchColor;
 		static bool TextureBarrier;
