@@ -35,30 +35,30 @@ bool WindowsWGL::start()
 		hWnd = GetActiveWindow();
 
 	if ((hDC = GetDC(hWnd)) == NULL) {
-		MessageBox(hWnd, L"Error while getting a device context!", pluginNameW, MB_ICONERROR | MB_OK);
+		MessageBoxW(hWnd, L"Error while getting a device context!", pluginNameW, MB_ICONERROR | MB_OK);
 		return false;
 	}
 
 	if ((pixelFormat = ChoosePixelFormat(hDC, &pfd)) == 0) {
-		MessageBox(hWnd, L"Unable to find a suitable pixel format!", pluginNameW, MB_ICONERROR | MB_OK);
+		MessageBoxW(hWnd, L"Unable to find a suitable pixel format!", pluginNameW, MB_ICONERROR | MB_OK);
 		stop();
 		return false;
 	}
 
 	if ((SetPixelFormat(hDC, pixelFormat, &pfd)) == FALSE) {
-		MessageBox(hWnd, L"Error while setting pixel format!", pluginNameW, MB_ICONERROR | MB_OK);
+		MessageBoxW(hWnd, L"Error while setting pixel format!", pluginNameW, MB_ICONERROR | MB_OK);
 		stop();
 		return false;
 	}
 
 	if ((hRC = wglCreateContext(hDC)) == NULL) {
-		MessageBox(hWnd, L"Error while creating OpenGL context!", pluginNameW, MB_ICONERROR | MB_OK);
+		MessageBoxW(hWnd, L"Error while creating OpenGL context!", pluginNameW, MB_ICONERROR | MB_OK);
 		stop();
 		return false;
 	}
 
 	if ((wglMakeCurrent(hDC, hRC)) == FALSE) {
-		MessageBox(hWnd, L"Error while making OpenGL context current!", pluginNameW, MB_ICONERROR | MB_OK);
+		MessageBoxW(hWnd, L"Error while making OpenGL context current!", pluginNameW, MB_ICONERROR | MB_OK);
 		stop();
 		return false;
 	}
@@ -102,7 +102,15 @@ bool WindowsWGL::start()
 
 		if (strstr(wglextensions, "WGL_EXT_swap_control") != nullptr) {
 			PFNWGLSWAPINTERVALEXTPROC wglSwapIntervalEXT = (PFNWGLSWAPINTERVALEXTPROC)wglGetProcAddress("wglSwapIntervalEXT");
-			wglSwapIntervalEXT(config.video.verticalSync);
+
+			// use adaptive vsync when supported and
+			// when vsync is enabled
+			if (strstr(wglextensions, "WGL_EXT_swap_control_tear") != nullptr &&
+				config.video.verticalSync > 0) {
+				wglSwapIntervalEXT(-1);
+			} else {
+				wglSwapIntervalEXT(config.video.verticalSync);
+			}
 		}
 	}
 
