@@ -8,6 +8,7 @@
 #include <Graphics/Context.h>
 #include <Graphics/Parameters.h>
 #include <DisplayWindow.h>
+#include <windows/ScreenShot.h>
 #include <Graphics/OpenGLContext/ThreadedOpenGl/opengl_Wrapper.h>
 
 using namespace opengl;
@@ -118,7 +119,7 @@ void DisplayWindowWindows::_changeWindow()
 		fullscreenMode.dmFields = DM_BITSPERPEL | DM_PELSWIDTH | DM_PELSHEIGHT | DM_DISPLAYFREQUENCY;
 
 		if (ChangeDisplaySettings( &fullscreenMode, CDS_FULLSCREEN ) != DISP_CHANGE_SUCCESSFUL) {
-			MessageBox( NULL, L"Failed to change display mode", pluginNameW, MB_ICONERROR | MB_OK );
+			MessageBoxW( NULL, L"Failed to change display mode", pluginNameW, MB_ICONERROR | MB_OK );
 			return;
 		}
 
@@ -166,6 +167,8 @@ bool DisplayWindowWindows::_resizeWindow()
 {
 	RECT windowRect, statusRect, toolRect;
 
+	windowRect = statusRect = toolRect = { 0 };
+
 	if (m_bFullscreen) {
 		m_screenWidth = config.video.fullscreenWidth;
 		m_screenHeight = config.video.fullscreenHeight;
@@ -179,12 +182,12 @@ bool DisplayWindowWindows::_resizeWindow()
 		_setBufferSize();
 
 		GetClientRect( hWnd, &windowRect );
-		GetWindowRect( hStatusBar, &statusRect );
+
+		if (hStatusBar)
+			GetWindowRect( hStatusBar, &statusRect );
 
 		if (hToolBar)
 			GetWindowRect( hToolBar, &toolRect );
-		else
-			toolRect.bottom = toolRect.top = 0;
 
 		m_heightOffset = (statusRect.bottom - statusRect.top);
 		windowRect.right = windowRect.left + config.video.windowedWidth - 1;
